requirements.txt
Django==2.1.7
django-cors-headers==2.4.0
django-pyodbc-azure==2.1.0.0
djangorestframework==3.9.1
itsdangerous==1.1.0
pyodbc==4.0.25
pytz==2018.9
